requirements.txt
#apache-airflow-providers-google==10.3.0
astronomer-cosmos[dbt-bigquery]==1.0.3
protobuf==3.20.0
setuptools